refactor: Refactor PartyInfo
Removed duplicate partyinfo from publication request sections, now only using from party sections, to prevent confusion later
