+++ src/components/PartySections/PartyInfo/PartyInfo.js
@@ -2,15 +2,105 @@ import PropTypes from 'prop-types';
 
 import { FormattedMessage } from 'react-intl';
 
-import { Col, KeyValue, Row } from '@folio/stripes/components';
+import {
+  Card,
+  Col,
+  KeyValue,
+  Label,
+  Row,
+} from '@folio/stripes/components';
 
 import ExternalLink from '../../ExternalLink';
 
 const propTypes = {
+  otherEmailAddresses: PropTypes.object,
   party: PropTypes.object,
+  streetAddresses: PropTypes.object,
 };
 
-const PartyInfo = ({ party }) => {
+const renderStreetAddresses = (streetAddresses) => {
+  return (
+    <>
+      {streetAddresses.map((address, index) => (
+        <Row>
+          <Card
+            headerStart={
+              <Label>
+                <FormattedMessage
+                  id="ui-oa.streetAddresses.streetAddressTitle"
+                  values={{ number: index }}
+                />
+              </Label>
+            }
+          >
+            <Row start="xs">
+              <Col xs={3}>
+                <KeyValue
+                  label={
+                    <FormattedMessage id="ui-oa.streetAddresses.addressType" />
+                  }
+                  value={address?.addressType}
+                />
+              </Col>
+              <Col xs={3}>
+                <KeyValue
+                  label={
+                    <FormattedMessage id="ui-oa.streetAddresses.addressLine1" />
+                  }
+                  value={address?.addressLine1}
+                />
+              </Col>
+              <Col xs={3}>
+                <KeyValue
+                  label={
+                    <FormattedMessage id="ui-oa.streetAddresses.addressLine1" />
+                  }
+                  value={address?.addressLine2}
+                />
+              </Col>
+              <Col xs={3}>
+                <KeyValue
+                  label={<FormattedMessage id="ui-oa.streetAddresses.city" />}
+                  value={address?.city}
+                />
+              </Col>
+            </Row>
+
+            <Row start="xs">
+              <Col xs={3}>
+                <KeyValue
+                  label={
+                    <FormattedMessage id="ui-oa.streetAddresses.stateRegion" />
+                  }
+                  value={address?.stateRegion}
+                />
+              </Col>
+              <Col xs={3}>
+                <KeyValue
+                  label={
+                    <FormattedMessage id="ui-oa.streetAddresses.postalCode" />
+                  }
+                  value={address?.postalCode}
+                />
+              </Col>
+              <Col xs={3}>
+                <KeyValue
+                  label={
+                    <FormattedMessage id="ui-oa.streetAddresses.country" />
+                  }
+                  value={address?.country}
+                />
+              </Col>
+              <Col xs={3} />
+            </Row>
+          </Card>
+        </Row>
+      ))}
+    </>
+  );
+};
+
+const PartyInfo = ({ party, streetAddresses, otherEmailAddresses }) => {
   return (
     <>
       <Row>
@@ -73,6 +163,37 @@ const PartyInfo = ({ party }) => {
           />
         </Col>
       </Row>
+
+      {otherEmailAddresses && (
+        <Row start="xs">
+          <Col xs={6}>
+            <KeyValue
+              label={
+                <FormattedMessage id="ui-oa.otherEmail.otherEmailAddresses" />
+              }
+            >
+              {otherEmailAddresses ? (
+                <ul>
+                  {otherEmailAddresses.map((email) => (
+                    <li key={email?.id}>{email?.emailAddress}</li>
+                  ))}
+                </ul>
+              ) : null}
+            </KeyValue>
+          </Col>
+        </Row>
+      )}
+
+      {streetAddresses && (
+        <Row>
+          <Col xs={12}>
+            <Label>
+              <FormattedMessage id="ui-oa.publicationRequest.streetAddresses" />
+            </Label>
+            {streetAddresses ? renderStreetAddresses(streetAddresses) : null}
+          </Col>
+        </Row>
+      )}
     </>
   );
 };
