+++ src/components/PublicationRequestFormSections/RequestContactForm/RequestContactForm.js
@@ -17,7 +17,7 @@ import {
 
 import { EditCard } from '@folio/stripes-erm-components';
 
-import PartyInfo from '../../PublicationRequestSections/PartyInfo';
+import PartyInfo from '../../PartySections/PartyInfo';
 import PartyModal from '../../PartyModal';
 import css from './RequestContactForm.css';
