+++ src/components/PublicationRequestSections/RequestContact/RequestContact.js
@@ -8,7 +8,7 @@ import {
   Badge,
 } from '@folio/stripes/components';
 
-import PartyInfo from '../PartyInfo';
+import PartyInfo from '../../PartySections/PartyInfo';
 
 const propTypes = {
   request: PropTypes.object
