+++ src/components/PublicationRequestSections/index.js
@@ -4,7 +4,6 @@ export { default as Publication } from './Publication';
 export { default as PublicationStatus } from './PublicationStatus';
 export { default as RequestContact } from './RequestContact';
 export { default as RequestInfo } from './RequestInfo';
-export { default as PartyInfo } from './PartyInfo';
 export { default as Correspondence } from './Correspondence';
 export { default as Agreement } from './Agreement';
